requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-google-genai
openai
pydantic
python-dotenv
requests
faiss-cpu
supabase
dotenv
fastmcp
mcp
aiohttp
flask-bcrypt
scikit-learn
numpy<2.0.0
pandas
nltk
sentence_transformers
torch
gspread
google-auth
google-auth-oauthlib
google-generativeai
PyJWT